requirements.txt
langchain>=0.1.10
langchain-community>=0.0.30
langchain-google-genai>=2.0.0
google-generativeai>=0.8.3
faiss-cpu
python-dotenv
pypdf